Extract _plex.py from _legacy.py
Moves the Plex query and metadata helpers into a dedicated submodule:

- get_top_imdb_items (top-25 audience-rated titles per library, IMDb-tagged)
- get_plex_key_by_name (dict-list lookup by name)
- _extract_imdb_id_from_item, _normalize_lookup_title (private helpers)
- find_item_by_title, find_item_by_imdb_id (Plex library search)
- get_plex_summary (formatted server summary)
- get_plex_maintenance_hours (butler start/end hours)
- get_library_summaries (formatted library summary)
- get_plex_metadata (cached full server + libraries payload)
- get_library_metadata (per-section agent/scanner/counts/ratings)

The plex cache re-integration uses the pre-existing lazy import pattern
(from modules import helpers as _h_plex_cache) — no behavioral change.

helpers/__init__.py adds 'from ._plex import *' so all public API
remains accessible via helpers.foo(). Blueprint call sites unchanged.

Reduces _legacy.py from 1005 to 635 lines (-370). All 1077 tests pass.
